+++ shipwright/build.py
@@ -13,7 +13,7 @@ def _merge(d1, d2):
     return d
 
 
-def do_build(client, build_ref, targets):
+def do_build(client, build_ref, targets, pull_cache):
     """
     Generic function for building multiple images while
     notifying a callback function with output produced.
@@ -39,11 +39,11 @@ def do_build(client, build_ref, targets):
         parent_ref = None
         if target.parent:
             parent_ref = build_index.get(target.parent)
-        for evt in build(client, parent_ref, target):
+        for evt in build(client, parent_ref, target, pull_cache):
             yield evt
 
 
-def build(client, parent_ref, image):
+def build(client, parent_ref, image, pull_cache):
     """
     builds the given image tagged with <build_ref> and ensures that
     it depends on it's parent if it's part of this build group (shares
@@ -62,7 +62,25 @@ def build(client, parent_ref, image):
 
     built_tags = docker.last_built_from_docker(client, image.name)
     if image.ref in built_tags:
-        return []
+        return
+
+    if pull_cache:
+        pull_evts = client.pull(
+            repository=image.name,
+            tag=image.ref,
+            stream=True,
+        )
+
+        failed = False
+        for evt in pull_evts:
+            event = process_event_(evt)
+            if 'error' in event:
+                failed = True
+            else:
+                yield event
+
+        if not failed:
+            return
 
     build_evts = client.build(
         fileobj=mkcontext(parent_ref, image.path),
@@ -73,4 +91,5 @@ def build(client, parent_ref, image):
         dockerfile=os.path.basename(image.path),
     )
 
-    return (process_event_(evt) for evt in build_evts)
+    for evt in build_evts:
+        yield process_event_(evt)