+++ tests/test_cli.py
@@ -16,7 +16,6 @@ def get_defaults():
         '--exclude': [],
         '--help': False,
         '--no-build': False,
-        '--dirty': False,
         '--upto': [],
         '--x-assert-hostname': False,
         '-H': None,
@@ -90,7 +89,6 @@ def test_args():
         '--exclude': [],
         '--help': False,
         '--no-build': False,
-        '--dirty': False,
         '--upto': [],
         '--x-assert-hostname': True,
         '-H': None,
@@ -105,7 +103,7 @@ def test_args_2():
     args = [
         '--account=x', '--x-assert-hostname', 'build',
         '-d', 'foo', 'bar',
-        '-t', 'foo', '--dirty',
+        '-t', 'foo', '--dirty', '--pull-cache',
     ]
     parser = cli.argparser()
     arguments = cli.old_style_arg_dict(parser.parse_args(args))
@@ -118,7 +116,6 @@ def test_args_2():
         '--exclude': [],
         '--help': False,
         '--no-build': False,
-        '--dirty': True,
         '--upto': [],
         '--x-assert-hostname': True,
         '-H': None,
@@ -142,7 +139,6 @@ def test_args_base():
         '--exclude': [],
         '--help': False,
         '--no-build': False,
-        '--dirty': False,
         '--upto': [],
         '--x-assert-hostname': False,
         '-H': None,